% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence_details.R
\name{occurrence_details}
\alias{occurrence_details}
\title{Retrieve the full details of occurrence records}
\usage{
occurrence_details(uuid, verbose = sbdi_config()$verbose)
}
\arguments{
\item{uuid}{string: one or more record ids, as returned by \code{\link{occurrences}} (in the \code{data$id} column)}

\item{verbose}{logical: show additional progress information? [default is set by sbdi_config()]}
}
\value{
A named list (named by uuid), each element of which is a list containing the details for that uuid. This inner list will be empty if no match is found for the supplied uuid
}
\description{
Note that this makes a separate web request for each occurrence uuid, and so may not be wise to use on a large number of uuids.
}
\examples{
\dontrun{
s1 <- occurrence_details("f31e5f0e-f964-4bc3-b8f3-78f2ad520563")
str(s1,max.level=3)
}
}
\references{
Associated SBDI web service: \url{https://api.biodiversitydata.se/#ws102}
}
\seealso{
\code{\link{occurrences}} \code{\link{sbdi_config}}
}
