% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbdi_cache_filename.R
\name{sbdi_cache_filename}
\alias{sbdi_cache_filename}
\title{Returns the name of the cache file associated with the given URL.}
\usage{
sbdi_cache_filename(url)
}
\arguments{
\item{url}{string: the URL e.g. \url{https://spatial.biodiversitydata.se/}}
}
\value{
string: the file path and name
}
\description{
Note that this file may not actually exist, this function just provides the mapping from URL to file name
}
\examples{
\dontrun{
sbdi_cache_filename("https://spatial.biodiversitydata.se/")
}

}
\references{
\url{https://api.biodiversitydata.se/}
}
\seealso{
\code{sbdi_config} for cache settings, particularly the cache directory
}
