% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe100kmGrid.R
\docType{data}
\name{swe100kmGrid}
\alias{swe100kmGrid}
\title{A hexagonal polygon grid over Sweden
A 100 km wide grid dataset in SpatialPolygons (package sp) over Sweden.}
\format{
SpatialPolygons (package sp) CRS: EPSG:4326
}
\source{
made with package BIRDS
}
\usage{
data(swe100kmGrid)
}
\description{
A hexagonal polygon grid over Sweden
A 100 km wide grid dataset in SpatialPolygons (package sp) over Sweden.
}
\keyword{datasets}
