% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sweden_Grid.R
\docType{data}
\name{Sweden_Grid}
\alias{Sweden_Grid}
\alias{Sweden_Grid_5km_Wgs84}
\alias{Sweden_Grid_10km_Wgs84}
\alias{Sweden_Grid_25km_Wgs84}
\alias{Sweden_Grid_50km_Wgs84}
\title{The standard Swedish squared grid}
\format{
SpatialPolygons (package sp) CRS: EPSG:4326
}
\source{
https://github.com/biodiversitydata-se/Sweden_Grids
}
\usage{
data(Sweden_Grid_5km_Wgs84)

data(Sweden_Grid_10km_Wgs84)

data(Sweden_Grid_25km_Wgs84)

data(Sweden_Grid_50km_Wgs84)
}
\description{
A 5, 10, 25 and 50 km wide grid dataset in SpatialPolygons (package sp) over Sweden.
}
\keyword{datasets}
