% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sweden_Grid_50km_Wgs84.R
\docType{data}
\name{Sweden_Grid_50km_Wgs84}
\alias{Sweden_Grid_50km_Wgs84}
\title{The standard Swedish squared grid 
A 50 km wide grid dataset in SpatialPolygons (package sp) over Sweden.}
\format{
SpatialPolygons (package sp) CRS: EPSG:4326
}
\source{
https://github.com/biodiversitydata-se/Sweden_Grids
}
\usage{
data(Sweden_Grid_50km_Wgs84)
}
\description{
The standard Swedish squared grid 
A 50 km wide grid dataset in SpatialPolygons (package sp) over Sweden.
}
\keyword{datasets}
