% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sweden_HexGrid_100km_Wgs84.R
\docType{data}
\name{Sweden_HexGrid_100km_Wgs84}
\alias{Sweden_HexGrid_100km_Wgs84}
\title{A hexagonal polygon grid over Sweden
A 100 km wide grid dataset in SpatialPolygons (package sp) over Sweden.}
\format{
SpatialPolygons (package sp) CRS: EPSG:4326
}
\source{
made with package BIRDS
}
\usage{
data(Sweden_HexGrid_100km_Wgs84)
}
\description{
A hexagonal polygon grid over Sweden
A 100 km wide grid dataset in SpatialPolygons (package sp) over Sweden.
}
\keyword{datasets}
